# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Soiree API
  version: 0.1.0
  description: |
    Soiree — Robocon test-run video manager HTTP API.

    Phase 1 の主要リソース（masters / sessions / videos / runs / markers / search）を契約として定義する。
    実装は spec-devflow.md §4〜§9 の順に handler / service / UI を縦断で積む。

    ## 共通方針

    - JSON フィールド名は camelCase
    - ID は UUID 文字列
    - 時刻は ISO 8601 (RFC 3339) タイムスタンプ
    - 動画上の時間は秒単位（int = 整数秒、float = 小数秒）
    - リスト系は `{ data: [...], pagination: {...} }` のエンベロープ
    - ページングは cursor ベース (`?cursor=...&limit=...`)
    - エラーは `ErrorResponse` で統一
servers:
  - url: http://localhost:8080
    description: Local development
  - url: /
    description: Same-origin (production / docker)
security:
  - xUserId: []
...
paths:
...
  /devices:
    get:
      tags: [devices]
      operationId: listDevices
      summary: 機材一覧
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceList"
    post:
      tags: [devices]
      operationId: createDevice
      summary: 機材作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeviceRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "400":
          $ref: "#/components/responses/BadRequest"
  /devices/{deviceId}:
    parameters:
      - $ref: "#/components/parameters/DeviceId"
    get:
      tags: [devices]
      operationId: getDevice
      summary: 機材取得
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [devices]
      operationId: updateDevice
      summary: 機材更新（時計オフセット含む）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDeviceRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [devices]
      operationId: deleteDevice
      summary: 機材削除
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------- Teams ----------
...
components:
  securitySchemes:
    xUserId:
      type: apiKey
      in: header
      name: X-User-Id
      description: |
        現在ユーザーの UUID。セルフホスト・全員同権限の前提のため
        Phase 1 はクライアントが選択ユーザーを header で送る簡易方式。
        将来正式な認証を入れたら securityScheme を差し替える。
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      description: 前回レスポンスの `pagination.nextCursor`
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      description: 1 レスポンスあたり最大件数（1〜200、既定 50）
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: リクエストが不正
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: リソースが存在しない
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: 機械可読なエラーコード
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    Pagination:
      type: object
      required: [hasMore]
      properties:
        hasMore:
          type: boolean
        nextCursor:
          type: [string, "null"]
          description: 次ページを取得するための cursor

    # ---------- Enums ----------
...
    Device:
      type: object
      required: [id, name, defaultTimeOffsetSec, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        defaultTimeOffsetSec:
          type: number
          description: 機材の時計ズレ（秒、小数可）。プラスなら機材時刻が進んでいる
        createdAt:
          type: string
          format: date-time
    CreateDeviceRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        defaultTimeOffsetSec:
          type: number
          default: 0
    UpdateDeviceRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        defaultTimeOffsetSec:
          type: number
    DeviceList:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Device"
        pagination:
          $ref: "#/components/schemas/Pagination"

    # ---------- Team ----------
...